Dynamic skin loading and unloading, improve skin list UI
Skin Loading and Unloading
--------------------------

Instead of loading all skins from the `skins` folder immediately, only scan the folder to determine which skins are available and lazily load skins when they are being used. Skins not present in the `skins` folder are loaded from `downloadedskins` or downloaded.

Load all skin data (image data, image data in grayscale, metrics, blood color) is background jobs, so in the main thread we only need to upload the skin data to the GPU and add the skin to the map.

Start background jobs to load vanilla skins (including special skins) immediately so this base set of skins is available directly. All other skins are only loaded on demand.

Load the `default` skin specifically on the main thread to ensure that it is immediately available as a placeholder for skins which are not loaded yet.

Add the config variable `cl_skins_loaded_max` to specify the maximum number of skins that can be loaded at the same time. This defaults to 256 skins on 32 bit and 512 skins on 64 bit.

Unload the least recently used skins when trying to load more skins than the maximum. This is implemented using a basic LRU cache with an `std::list`, which is much more efficient then searching through the map of all skins when there are many skins (e.g. with the entire skin database).

Skin List UI Improvements
-------------------------

Show progress indicator for skins that are loading.

Show error icon with tooltip for skins that failed to be loaded due to a file/download error.

Show error icon with tooltip when entering an invalid or special skin name.

Show question mark icon with tooltip for skins that could not be found locally or in the skin database.

Show error message and button to reset the filter when the skin filter does not match any available skin.

Increase the scrolling speed so 2 rows of skins are scrolled with the mouse wheel to make the scrolling speed more consistent with other lists.

Cache information about the skin list to avoid duplicate calculations and improve the performance of skin list rendering.

Show how many skins are in which state when `debug` is enabled.