snap: exclude team score and flag carrier from snap if not needed
I tested it in game both 0.6 and 0.7 would support receiving all this
data at all times and nothing breaks.

But if the gamemode did not set the flag (ctf) or team gameflag we do
not need to send it. So this commit makes for example the "dm" mode
snapshots more lightweight.