version: downgrade from v2.0-dev to v1.7+dev
It was a cute idea to plan with a 2.0 release.
But planning release numbers in advance does not work
well for me. So instead of predicting the next version
number and saying its a pre release of it.
We swap it around and keep the old release and append the
"+dev" suffix. Meaning its a postrelease.

So the current latest version is the latest stable release
plus new development patches.

Thus we are back to `v1.7+dev`

The plan for v2.0 was database breaking changes.
And these two big prs that would increase bugs and maintenance:

- #345 (now closed)
- #264 (old draft and it also doesn't look like its going anywhere)

So the next release will probably nothing ground breaking.
But just a new stable version with all the new tiny fixes
and addition of new features.